third_party/blink/web_tests/platform/generic/http/tests/devtools/tracing/timeline-time/timeline-usertiming-expected.txt

Test performance.mark/measure records on Timeline


Running: testSimplePerformanceMeasure
a

Running: testNestedPerformanceMeasure
a
b
c
d

Running: testUnbalancedPerformanceMeasure
a
b

Running: testUnnestedPerformanceMeasure
a
b

Running: testParentMeasureIsOnTop
durationTimeTotal
durationTime1
durationTime2
